+++ tests/fixtures/index.html
@@ -10,13 +10,27 @@
     <title>Document</title>
   </head>
 
-  <body class="font-sans">
-    <h1 class="text-lg leading-none capsize mt-6">Heading 1</h1>
+  <body class="font-sans space-y-6">
+    <h1 class="text-lg leading-none capsize">Heading 1</h1>
 
-    <h2 class="text-[20] leading-normal capsize mt-6">
+    <h2 class="text-[20] leading-normal capsize">
       Sed vero dolores no ut diam dolores gubergren duis sed accumsan kasd lorem
       clita et iriure ipsum tempor. Dolore eos ut eum placerat labore sadipscing
       tation dolores voluptua iusto tempor dolor labore feugiat stet.
     </h2>
+
+    <div class="text-lg leading-normal space-y-6">
+      <p class="capsize">
+        Sadipscing dolore ea nihil labore consequat clita enim ipsum commodo
+        sanctus dolor rebum dolor. Sed possim in ex rebum magna kasd amet et
+        molestie ex augue et nam.
+      </p>
+
+      <p class="capsize">
+        Sadipscing dolore ea nihil labore consequat clita enim ipsum commodo
+        sanctus dolor rebum dolor. Sed possim in ex rebum magna kasd amet et
+        molestie ex augue et nam.
+      </p>
+    </div>
   </body>
 </html>
