--- tests/index.test.ts
@@ -1,4 +1,0 @@
-import { expect, test, vi } from 'vitest'
-import { runTailwind } from './_testUtils'
-
-test('works', () => {})
